% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zmien_obiekty_skalowania.R
\name{daj_id_skali}
\alias{daj_id_skali}
\title{Pobiernie id skali}
\usage{
daj_id_skali(rokEgzaminu, rodzajEgzaminu, skrot = NULL,
  zrodloDanychODBC = "EWD_grzes")
}
\arguments{
\item{rokEgzaminu}{rok egzaminu powiązany ze skalą}

\item{rodzajEgzaminu}{rodzaj egzaminu"}

\item{skrot}{skrótowa nazwa egzaminu lub jego części (np: "s", "gh" lub "m_mp")}

\item{zrodloDanychODBC}{opcjonalnie nazwa źródła danych ODBC, dającego dostęp do bazy
(domyślnie "EWD")}
}
\value{
typ numeryczny
}
\description{
Funkcja na podstawie ogulnych informacji o egzaminie pobiera jego id skali.
}

