% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{.e}
\alias{.e}
\title{Wyparsowuje znaki konca stringu}
\usage{
.e(str)
}
\arguments{
\item{str}{lancuch znakow do wyparsowania}
}
\value{
character
}
\description{
Funkcja do wyparsowywania niebezpiecznych znakow przy budowaniu zapytan SQL
}
\details{
_
}
