% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{.sqlPlaceholders}
\alias{.sqlPlaceholders}
\title{Zwraca placeholdery parametrów zapytania SQL}
\usage{
.sqlPlaceholders(dane, start = 1)
}
\arguments{
\item{dane}{cokolwiek, na czym zadziała \code{length()}}

\item{start}{numer pierwszego placeholdera}
}
\value{
character
}
\description{
_
}
