% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{.sqlQuery}
\alias{.sqlQuery}
\title{Wykonuje zapytanie sql i obsluguje bledy}
\usage{
.sqlQuery(P, sql, dane = NULL)
}
\arguments{
\item{P}{otwarte polaczenie ODBC}

\item{sql}{polecenie SQL do wykonania}

\item{dane}{ramka danych z danymi, które mają zostać wykorzystane w zapytaniu}
}
\value{
data.frame
}
\description{
_
}
\details{
_
}
