% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zapisz_skalowanie.R
\name{przygotuj_kryteria_do_edycji_skali}
\alias{przygotuj_kryteria_do_edycji_skali}
\title{Edycja skali}
\usage{
przygotuj_kryteria_do_edycji_skali(skalowanie)
}
\arguments{
\item{skalowanie}{wynik działania fukcji skaluj z dodanym parametrem id_skali}
}
\value{
data frame w formacie parametru funkcji \code{\link{edytuj_skale}}
}
\description{
Zapis skalowania do bazy danych
}

