% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zmien_obiekty_skalowania.R
\name{przypisz_grupy_do_oszacowan}
\alias{przypisz_grupy_do_oszacowan}
\title{Dodanie grupy do oszacowań umiejętności}
\usage{
przypisz_grupy_do_oszacowan(skalowanieWynik, rodzaj = NULL,
  maskaTestowa = NULL)
}
\arguments{
\item{skalowanieWynik}{lista opisująca wynik pojedynczego skalowania}

\item{rodzaj}{nazwa równa "LO" lub "T" dla matury albo domyśnie NULL dla sprawdzianu i egzaminu gimnazjalnego}

\item{maskaTestowa}{wektor całkowitoliczbowy określający, które obserwacje mają być umieszczone w docelowym obiekcie.
Jeżeli NULL to wszystkie obserwacje są umieszczane}
}
\value{
data frame zawierający umiejętności z dodaną kolumną grupy
}
\description{
Funkcja dodaje kolumnę grupa do elementu listy skalowania opisującego umiejętności uczniów
}

