% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zmien_obiekty_skalowania.R
\name{przypisz_grupy_do_parametrow}
\alias{przypisz_grupy_do_parametrow}
\title{Dodanie grupy do parametrów zadań}
\usage{
przypisz_grupy_do_parametrow(skalowanieWynik, grupy)
}
\arguments{
\item{skalowanieWynik}{lista opisująca wynik pojedynczego skalowania}

\item{grupy}{wektor stringów - nazwy grup, które należy dodać}
}
\value{
data frame zawierający parametry zadań z dodaną kolumną opisującą grupę
}
\description{
Funkcja dodaje kolumnę grupa do elementu listy skalowania opisującego parametry zadań
}

