% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/zapisz_parametry_skalowania.R
\name{sprawdz_zgodnosc_kryteriow}
\alias{sprawdz_zgodnosc_kryteriow}
\title{Sprawdzanie zgodnosci kryteriow.}
\usage{
sprawdz_zgodnosc_kryteriow(kryt1, kryt2, nazwa1, nazwa2, error = TRUE)
}
\arguments{
\item{kryt1}{wektor liczb opisujący kryteria}

\item{kryt2}{wektor liczb opisujący kryteria}

\item{nazwa1}{nazwa opisująca wektor kryt1}

\item{nazwa2}{nazwa opisująca wektor kryt2}

\item{error}{jeżeli TRUE to funkcje wyświetla błąd w przypadku, gdy któryś z elementów kryt1
nie należy do kryt2.}
}
\value{
Funkcja nie zwraca żadnej wartości.
}
\description{
Funkcja sprawdza, czy wektor kryteriów kryt1 jest zawarty w zbiorze kryteriów kryt2.
}

