% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stworz_nowe_wskazniki_ewd.R
\name{stworz_nowe_wskazniki_ewd}
\alias{stworz_nowe_wskazniki_ewd}
\title{Zapis wskaznikow EWD do bazy}
\usage{
stworz_nowe_wskazniki_ewd(wskaznikiWzorce, wyrazeniaZmienWskaznik,
  wyrazeniaZmienNazwy = NULL, wyrazeniaZmienOpisy = NULL, okres = NULL,
  doPrezentacji = FALSE, zrodloDanychODBC = "ewd")
}
\arguments{
\item{wskaznikiWzorce}{wektor ciągów znaków - identyfikatory wskaźników,
które już istnieją w bazie i mają posłużyć za wzorce dla nowotworzonych;}

\item{wyrazeniaZmienWskaznik}{dwuelementowy wektor ciągów znaków: pierwszy
element zostanie podany jako argument \code{pattern}, a drugi jako argument
\code{replace} do funkcji \code{\link{sub}} wywołanej na wartościach
argumentu \code{wyrazeniaZmienWskaznik}, w celu utworzenia identyfikatorów
nowych wskaźników}

\item{wyrazeniaZmienNazwy}{opcjonalnie dwuelementowy wektor ciągów znaków:
pierwszy element zostanie podany jako argument \code{pattern}, a drugi jako
argument \code{replace} do funkcji \code{\link{sub}} wywołanej na nazwach
"wzorcowych" wskaźników, w celu uzyskania nazw nowych wskaźników}

\item{wyrazeniaZmienOpisy}{opcjonalnie dwuelementowy wektor ciągów znaków:
pierwszy element zostanie podany jako argument \code{pattern}, a drugi jako
argument \code{replace} do funkcji \code{\link{sub}} wywołanej na opisach
"wzorcowych" wskaźników, w celu uzyskania opisów nowych wskaźników}

\item{okres}{liczba całkowita, zostanie przypisana nowym wskaźnikom
w kolumnie \code{okres} tablicy \code{sl_wskazniki}}

\item{doPrezentacji}{wartość logiczna, zostanie przypisana nowym wskaźnikom
w kolumnie \code{do_prezentacji} tablicy \code{sl_wskazniki}}

\item{zrodloDanychODBC}{opcjonalnie ciąg znaków - nazwa źródła danych ODBC,
dającego dostęp do bazy (domyślnie "ewd")}
}
\value{
funkcja nic nie zwraca
}
\description{
Funkcja pozwala utworzyć w bazie (w tablicy \code{sl_wskazniki}) nowe
wskaźniki, na podobieństwo już istniejących. Z podanych wskaźników skopiowane
zostaną wszystkie ich właściwości (w tym wpisy w tablicach
\code{sl_wskazniki_typy_szkol} i \code{sl_kategorie_lu}), a zmienione
zostaną identyfikatory wskaźników (tj. wartość kolumny \code{wskaznik}).
Dodatkowe argumenty pozwalają też zmodyfikować wartości kolumn \code{nazwa},
\code{opis}, \code{okres} i \code{do_prezentacji}.
}

