% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stworz_skale.R
\name{stworz_skale}
\alias{stworz_skale}
\title{Tworzy w bazie nowa skale i zwraca jej id_skali}
\usage{
stworz_skale(P, opis, rodzaj, doPrezentacji, idTestow, pominTransakcje)
}
\arguments{
\item{P}{połączenie z bazą danych uzyskane z \code{DBI::dbConnect(RPostgres::Postgres())}}

\item{opis}{opis skali (zwyczajowo w formacie "ZESPOL;EGZ;LATA", np. "paou;s;2002-2011" lub "ewd;gh_h;2012")}

\item{rodzaj}{rodzaj skali (ewd/zrównywanie/ktt)}

\item{doPrezentacji}{czy skala ma być oznaczona jako przeznaczona do prezentacji}

\item{idTestow}{wektor id testów, z którymi ma być powiązana skala}

\item{pominTransakcje}{wartość logiczna pozwalająca wywołać funkcję tak, aby nie używała
transakcji - co do zasady nie należy stosować (przydatne tylko, jako sposób na uniknięcie
błędu DBI/RPostgres związanego z brakiem obsługi zagnieżdżonych transakcji)}
}
\value{
[numeric] id_skali utworzonej skali
}
\description{
Funkcja stworz_skale() służy do „zarejestrowania” w bazie nowej skali.
Po dokonaniu takiego „zarejestrowania” należy wprowadzić do bazy informacje
o kryteriach oceny i pseudokryteriach oceny, z których składa się dana skala
- służy do tego funkcja \code{\link{edytuj_skale}}.
}
