% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stworz_test_z_wielu_czesci.R
\name{stworz_test_z_wielu_czesci}
\alias{stworz_test_z_wielu_czesci}
\title{Tworzenie testów będących złączeniem kilku części egzaminu.}
\usage{
stworz_test_z_wielu_czesci(rodzajEgzaminu, czesciEgzaminu, rokEgzaminu, czyEwd,
  opis, zrodloDanychODBC = "EWD")
}
\arguments{
\item{rodzajEgzaminu}{rodzaj egzaminu ("sprawdzian" / "egzamin gimnazjalny" / "matura")}

\item{czesciEgzaminu}{wektor części egzaminu, które mają zostać połączone}

\item{rokEgzaminu}{rok egzaminu}

\item{czyEwd}{czy dane mają pochodzić z testów EWD}

\item{opis}{opis nowetworzonego testu w tablic \code{testy}}

\item{zrodloDanychODBC}{opcjonalnie nazwa źródła danych ODBC, dającego dostęp do bazy
(domyślnie "EWD")}
}
\value{
id_testu utworzonego testu
}
\description{
Funkcja tworzy nowy test ze wskazanych części wskazanego (poprzez rodzaj i rok)
egzaminu i wypełnia tablicę \code{dane_osobowe.testy_obserwacje} uczniami, którzy
wystąpili w przynajmniej jednej z części wskazanego egzaminu.
}
\details{
Tworzony test posiada datę, jak najstarsza z data części egzaminu, z których powstał
oraz \code{ewd} zgodne z testami, z których powstał.

Spośród kolumn tablicy \code{dane_osobowe.testy_obserwacje} kopiowana jest tylko
kolumna \code{id_szkoly}, natomiast kolumna \code{zrodlo} ustawiana jest na wartość
\code{baza}. Pozostałe kolumny (np. \code{dysleksja, klasa}, itp.) ustawiane są na
\code{NULL}.

Jeśli w bazie istnieje już test o takim samym opisie, zrówcone zostanie id_testu tego
testu i jednocześnie wygenerowany zostanie warning informujący o tym, że nowy test nie
został utworzony. Uwaga! Nie gwarantuje to, że istniejący w bazie test składa się
z identycznych testów składowych - porównanie bazuje tylko na opisie testu, który jest
ustawiany arbitralnie przez tworządego test.
}

