% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zapisz_skalowanie.R
\name{uloz_insert_z_ramki}
\alias{uloz_insert_z_ramki}
\title{Zapis wynikow skalowania do bazy}
\usage{
uloz_insert_z_ramki(nazwa, ramka)
}
\arguments{
\item{nazwa}{ciąg znaków - nazwa tablicy w bazie, do której dane mają być
zapisywane}

\item{ramka}{data frame, którego struktura odpowiada strukturze tablicy
podanej argumentem \code{nazwa}}
}
\value{
ciąg znaków
}
\description{
Funkcja przygotowuje składnię polecenie INSERT dla zapytania
parametryzowanego, na podstawie struktury ramki danych.
}

