% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/usun_kryteria_ze_skali.R
\name{usun_kryteria_ze_skali}
\alias{usun_kryteria_ze_skali}
\title{Usuwanie zbędnych kryteriów i pseudokryteriów ze skali}
\usage{
usun_kryteria_ze_skali(skalowanie, idSkali, zrodloDanychODBC = "EWD")
}
\arguments{
\item{skalowanie}{lista zawierająca obiekt usunieteKryteria - wektor nazw kryteriów do
usunięcia.}

\item{idSkali}{id skali, z której usuwane są kryteria.}

\item{zrodloDanychODBC}{opcjonalnie nazwa źródła danych ODBC, dającego dostęp do bazy
(domyślnie "EWD")}
}
\value{
Funkcja zwraca ramkę danych, która podana do funkcji
\code{\link[ZPDzapis]{edytuj_skale}} usunie z niej zbędne (pseudo)kryteria.
}
\description{
Funkcja przygotowuje obiekt do modyfikacji skali, który nie zawiera zbędnych kryteriów
i pseudokryteriów.
}

