% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zapisz_parametry_skalowania.R
\name{wydziel_kryteria_pseudokryteria}
\alias{wydziel_kryteria_pseudokryteria}
\title{Wydzielanie kryteriow i pseudokryteriow}
\usage{
wydziel_kryteria_pseudokryteria(nazwy)
}
\arguments{
\item{nazwy}{wektor ciągów znakowych}
}
\value{
Funkcja zwraca dwuelementową listę. Pierwszy element o nazwie kryteria zawiera wektor liczb,
który zawiera numery kryteriow. Drugi element to wektor numerów pseudokryteriów.
}
\description{
Funkcja wydziala kryteria i pseudokryteria z wektora ciągów znakowych oraz sprawdza poprawność ich nazw.
}

