% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zapisz_normy.R
\name{zapisz_normy}
\alias{zapisz_normy}
\title{Zapis normalizacji ekwikwantylowej do bazy}
\usage{
zapisz_normy(P, normy, prefiksRok)
}
\arguments{
\item{P}{połączenie z bazą danych uzyskane z \code{DBI::dbConnect(RPostgres::Postgres())}}

\item{normy}{wektor opisujący unormowanie, typowo zwrócony przez funkcję
\code{\link[ZPD]{normalizuj}}}

\item{prefiksRok}{ciąg znaków postaci \code{'pr'}, gdzie \code{p} oznacza
prefiks części egzaminu, a \code{r} rok (zapis czterema cyframi)}
}
\value{
funkcja nic nie zwraca
}
\description{
Funkcja służy do zapisania do bazy obliczonych norm ekwikwantylowych.
}
\details{
Funkcja nie dopuszcza braków danych w argumencie \code{normy}.
Funkcja wymaga też, aby w ramach połączenia nawiązywanego z bazą mieć prawa do
modyfikacji tablic 'skale' i 'normy_ekwikwantylowe'.
}
\seealso{
\code{\link[ZPD]{normalizuj}}
}
