% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zapisz_oszacowania_umiejetnosci.R
\name{zapisz_oszacowania_umiejetnosci}
\alias{zapisz_oszacowania_umiejetnosci}
\title{Zapisuje oszacowania umiejętności do bazy danych}
\usage{
zapisz_oszacowania_umiejetnosci(oszacowania, skrotCzesci, rEAP, rodzajEgzaminu,
  rokEgzaminu, nrSkalowania, idSkali, rodzajEstymacji = "EAP",
  zrodloDanychODBC = "EWD", maskaTestowa = NULL)
}
\arguments{
\item{oszacowania}{ramka danych zawierająca oszacowania, błędy standardowe oraz id
testu lub testów poszczególnych uczniów.}

\item{skrotCzesci}{skrócona nazwa częsci egzaminu, np.: gh_p, gh.}

\item{rEAP}{rzetelność empiryczna.}

\item{rodzajEgzaminu}{rodzaj egzaminu.}

\item{rokEgzaminu}{rok egzaminu.}

\item{nrSkalowania}{numer skalowania dla skali.}

\item{idSkali}{id skali, z której usuwane są kryteria.}

\item{rodzajEstymacji}{ciąg znakowy opisujący rodzaj estymacji. Domyślnie przyjmuje
wartość 'EAP'.}

\item{zrodloDanychODBC}{opcjonalnie nazwa źródła danych ODBC, dającego dostęp do bazy
(domyślnie "EWD")}

\item{maskaTestowa}{wektor liczb określający, które wiersze ramki 'oszacowania' mają
być zapisane.}
}
\value{
Funkcja zwraca ramkę danych, która podana do funkcji
\code{\link[ZPDzapis]{edytuj_skale}} usunie z niej zbędne (pseudo)kryteria.
}
\description{
Funkcja zapisuje oszacowania umiejętności uczniów do bazy danych.
}

