% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zapisz_wskazniki_ewd.R
\name{zapisz_wskazniki_ewd}
\alias{zapisz_wskazniki_ewd}
\title{Zapis wskaznikow EWD do bazy}
\usage{
zapisz_wskazniki_ewd(nazwaPliku, sufiks, doPrezentacji = FALSE,
  nadpisz = FALSE, zrodloDanychODBC = "ewd")
}
\arguments{
\item{nazwaPliku}{ciąg znaków - nazwa pliku .RData zawierającego obiekt
klasy \code{listaWskaznikowEWD}, będący wynikiem działania funkcji
\code{\link[EWDwskazniki]{przygotuj_wsk_ewd}})}

\item{sufiks}{ciąg znaków - sufiks określający typ wskaźnika, dopisywany
do części wynikającej z nazwy konstruktu}

\item{doPrezentacji}{wartość logiczna - czy oznaczyć zapisywany wskaźnik
(w danym roku) jako 'do prezentacji'?}

\item{nadpisz}{wartość logiczna - czy nadpisać wartości zapisane w bazie?
(jeśli istnieją)}

\item{zrodloDanychODBC}{opcjonalnie ciąg znaków - nazwa źródła danych ODBC,
dającego dostęp do bazy (domyślnie "ewd")}
}
\value{
funkcja nic nie zwraca
}
\description{
Funkcja zapisuje do bazy oszacowania wskaźników EWD: wartości samych
wskaźników, liczbę zdających, parametry warstwic, wartości parametrów modeli
regresji oraz informacje o powiązaniu ze skalami wyników (oszacowań
umiejętności uczniów).
}
\details{
Wskaźniki, których wartości mają być wczytywane do bazy muszą już być
wcześniej zarejestrowane w tablicach \code{sl_wskazniki},
\code{sl_wskazniki_typy_szkol} i \code{sl_kategorie_lu}. Pomocna może być
w tym funkcja \code{\link{stworz_nowe_wskazniki_ewd}}

Parametr \code{doPrezentacji} wpływa na ustawienia na poziomie wskaźnika-roku,
ale o tym, czy wskaźniki będą publicznie dostępne decyduje koniunkcja
z wartością kolumny \code{do_prezentacji} w tablicy \code{sl_wskazniki},
na którą wywołanie tej funkcji nie oddziałuje (trzeba ją modyfikować w bazie
"ręcznie").
}

