% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zmien_obiekty_skalowania.R
\name{zmien_obiekty_skalowania}
\alias{zmien_obiekty_skalowania}
\title{Dodanie do obiektu skalowania grupy}
\usage{
zmien_obiekty_skalowania(skalowanie, rokEgzaminu, rodzajEgzaminu,
  zrodloDanychODBC = "EWD_grzes", maskaTestowa = NULL)
}
\arguments{
\item{skalowanie}{lista opisująca wynik skalowania}

\item{rokEgzaminu}{rok egzaminu}

\item{rodzajEgzaminu}{rodzaj egzaminu ("sprawdzian", "egzamin gimnazjalny", "matura")}

\item{zrodloDanychODBC}{opcjonalnie nazwa źródła danych ODBC, dającego dostęp do bazy
(domyślnie "EWD")}

\item{maskaTestowa}{wektor całkowitoliczbowy określający, które obserwacje mają być umieszczone w docelowym obiekcie.
Jeżeli NULL to wszystkie obserwacje są umieszczane}
}
\value{
data frame zawierający parametry zadań z dodaną kolumną opisującą grupę
}
\description{
Funkcja dodaje kolumnę grupa oraz id skali do odpowiednich elementów listy - wyniku skalowania.
}

