% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/znajdz_pseudokryteria.R
\name{znajdz_pseudokryteria}
\alias{znajdz_pseudokryteria}
\title{Odnajduje i tworzy pseudokryteria na podstawie list ich elementow}
\usage{
znajdz_pseudokryteria(P, kryteria, opisy)
}
\arguments{
\item{P}{połączenie z bazą danych uzyskane z \code{DBI::dbConnect(RPostgres::Postgres())}}

\item{kryteria}{macierz id_kryterium - kazdy wiersz opisuje jedno pseudokryterium}

\item{opisy}{wektor opisow pseudokryteriow}
}
\value{
[numeric vector] wektor odnalezionych i/lub utworzonych id_pseudokryteriow
}
\description{
_
}
