% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zapisz_parametry_skalowania.R
\name{zapisz_parametry_skalowania}
\alias{zapisz_parametry_skalowania}
\title{Zapisywanie parametrow skalowan.}
\usage{
zapisz_parametry_skalowania(nazwa_skali = NULL, id_testu = NULL,
  doPrezentacji, parametry, opis, estymacja, rEAP = NULL,
  zrodloDanychODBC = "EWD")
}
\arguments{
\item{nazwa_skali}{nazwa skali. Kiedy nazwa skali przymuje wartość NULL nie jest
brana pod uwagę w wyszukiwaniu. Wartość domyślna to NULL.}

\item{id_testu}{id testu. Kiedy id testu przymuje wartość NULL nie jest brana pod
uwagę w wyszukiwaniu. Wartość domyślna to NULL.}

\item{doPrezentacji}{wartość boolowska określająca, czy wstawiane skalowanie ma być do prezentacji}

\item{parametry}{ramka danych o strukturze zgodnej z tą, w jakiej zwraca oszacowania parametrów funkcja skaluj().}

\item{opis}{opis estymacji}

\item{estymacja}{parametr opisujący metodę estymacji - musi być jedną z wartości
występujących w tablicy sl_estymacje_parametrow.}

\item{rEAP}{rzetelność empiryczna, która domyśnie wynosi NULL.}

\item{zrodloDanychODBC}{string określający źródło danych. Wartość domyślna to 'EWD'.}
}
\value{
Funkcja zwraca id skalowania.
}
\description{
Funkcja zapisuje parametry skalowania do bazy danych
}

