% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zapisz_skalowanie.R
\name{zapisz_skalowanie}
\alias{zapisz_skalowanie}
\title{Zapisz obiekt skalowania}
\usage{
zapisz_skalowanie(skalowanie, opisSkalowania, nadpisz = FALSE,
  zrodloDanychODBC = "EWD_grzes")
}
\arguments{
\item{skalowanie}{wynik działania fukcji skaluj z dodanym parametrem id_skali}

\item{opisSkalowania}{lista zawierająca dane skalowania}

\item{nadpisz}{czy elementy skali powinny być nadpisane}

\item{zrodloDanychODBC}{opcjonalnie nazwa źródła danych ODBC, dającego dostęp do bazy
(domyślnie "EWD")}
}
\value{
funkcja nic nie zwraca
}
\description{
Zapis skalowania do bazy danych
}

